requirements.txt
# Framework Dependencies
pydantic>=2.0.0
pyyaml>=6.0
numpy>=1.20.0

# LLM Clients
openai>=1.0.0
anthropic>=0.5.0

# Environment Wrappers
stulife-campus>=2.0.0  # StuLife environment
alfworld>=0.3.0,<1.0.0  # ALFWorld environment (0.3.x for faster reset)
webshop>=1.0.0           # WebShop environment

# Development
pytest>=7.0.0
pytest-asyncio>=0.21.0
pytest-cov>=4.0.0
black>=22.0.0

# Async Support (for Agent runtime)
aiofiles>=23.0.0
aiosqlite>=0.19.0
aiohttp>=3.9.0
